Add files via upload
MorseCode Game with 2 pages. The first page explains what morse code is and there is a section at the bottom of the first page where the user can type in a letter or a word and it appears in morse code.
Then the second page consists of a practice game. The user must translate morse code by typing in the translation with letters of the Alphabet